some more refactors
