+++ app.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
 pip install -r requirements.txt --user
-python dash-cost-dashboard.py
+python dash_cost_dashboard.py
