+++ domino_cost/constants.py
@@ -1,0 +1,7 @@
+NO_TAG = "No tag"
+
+window_to_param = {
+    "30d": "Last 30 days",
+    "14d": "Last 14 days",
+    "7d": "Last 7 days"
+}
