+++ requirements.txt
@@ -20,3 +20,6 @@ dash-table-experiments~=0.6
 dash-bootstrap-components~=1.6
 pandas~=2.2
 pytest~=7.4.3
+
+requests~=2.32.3
+plotly~=5.23.0
